Split slow tests into separate files
